backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Config & validation
pydantic-settings==2.2.1
python-dotenv==1.0.1

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
python-multipart==0.0.9

# AWS
boto3==1.34.101

# AI
openai==1.30.1
